+++ dockerfile/php-fpm/CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 
-* resupport 5.6 7.0
-* add nightly, build from php-src git
-* Docker build enable buildkit
+* Please use `$ docker buildx build xxx` build image, Docker 19.03+
+* Resupport `5.6` `7.0`
+* Add `nightly-TYPE-alpine`, build from php-src git
+* Support buildkit